refactor(git): relocate github-commits signal config out of list framework
